文件目录、blob、vle修改
