+++ package.json
@@ -28,6 +28,7 @@
     "gsap": "^3.12.2",
     "mdast-util-to-string": "^3.2.0",
     "mdx-annotations": "^0.1.1",
+    "mermaid": "^10.3.0",
     "next": "13.4.2",
     "postcss-focus-visible": "^6.0.4",
     "react": "18.2.0",
