+++ src/components/git/variable-length/coding/Coding.module.css
@@ -1,4 +1,4 @@
-.VLC {
+.Coding {
     border: 1px solid darkseagreen;
     padding: 18px;
     height: 400px;
