+++ src/pages/blob.mdx
@@ -1,9 +1,14 @@
-import Blob from '../components/Blob/Blob.jsx';
+import Blob from '../components/git/blob/Blob.jsx';
 
 export const description =
   ''
 
 # Blob
+
+```
+这里中文算两位，英文算一位，可以中英文混合输入
+```
+
 <div>
     <Blob/>
 </div>
