+++ src/pages/index.mdx
@@ -8,6 +8,8 @@ export const sections = [
   { title: 'Objects', id: 'objects' },
 ]
 
+
+
 <HeroPattern />
 
 # Git - From Intern(al) To Expert
