+++ src/pages/variable-length.mdx
@@ -1,5 +1,5 @@
-import VLC from '../components/VLC/VLC.jsx';
-import VLD from '../components/VLD/VLD.jsx'
+import Coding from '../components/git/variable-length/Coding/Coding.jsx';
+import Decoding from '../components/git/variable-length//Decoding/Decoding.jsx'
 export const description =
   ''
 
@@ -12,11 +12,11 @@ export const description =
 来不为空）也可能是格式冲突，这个问题还待解决
 ```
 <div>
-    <VLC/>
+    <Coding/>
 </div>
 
 ## 解码
 
 <div>
-    <VLD/>
+    <Decoding/>
 </div>
